Compute the Likelihood of the BiSSeCo model

Likelihood of the BiSSeCo model given rates of variant mutation, selection, a dated tree, and effective population size over time

Usage

loglikelihood_bisseco(
  parms,
  mu,
  gamma,
  dtr,
  Net,
  augment_likelihood = TRUE,
  res = 200,
  ...
)

Arguments

parms

a 3-vector containing parameters alpha (ratio substitution rate to variant relative to ancestral type), omega (relative fitness = 1+s) and yscale (population size adjustment)

mu

Mean clock rate of evolution

gamma

1 / generation time

dtr

A phydynR::DatedTree

Net

Effective population size through time. This is stored as a two column

augment_likelihood

if TRUE (default), will combine the coalescent likelihood with a binomial likelihood of sampling variants or ancestral types under the assumption of random sampling and mutation-selection balance

res

integer number of time points used for coalescent approximation matrix, such that the first column is time and the second column is population size

...

Additional parameters are passed to colik